requirements.txt
fastapi
uvicorn
bcrypt
python-dotenv
sqlalchemy